+++ test/test_restconf_yang_patch_json.sh
@@ -9,8 +9,8 @@ s="$_" ; . ./lib.sh || if [ "$s" = $0 ]; then exit 0; else return 0; fi
 
 # Enable if YANG_PATCH
 echo "...skipped: YANG_PATCH JSON NYI"
-if [ "$s" = $0 ]; then exit 0; else return 0; fi
-    
+if [ -z "${CLIXON_YANG_PATCH}" -a "$s" = $0 ]; then exit 0; else return 0; fi
+
 APPNAME=example
     
 cfg=$dir/conf.xml